requirements.txt
asgiref==3.5.2
cloudinary==1.30.0
crispy-bootstrap5==0.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.16
django-allauth==0.51.0
django-bootstrap-datepicker-plus==5.0.2
django-braces==1.15.0
django-crispy-forms==1.14.0
django-model-utils==4.3.1
django-star-ratings==0.9.2
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests-oauthlib==1.3.1
sqlparse==0.4.3
swapper==1.3.0
